backend/requirements.txt
Flask
Flask-RESTful
Flask-JWT-extended
Flask-SQLAlchemy
cryptography
flask-socketio
pycryptodome
sqlalchemy-utils